refactor(shared): split webview IPC schema monoliths into per-domain modules
mainView and progressView become folders (state/inbound/outbound/data with
an index barrel at the original import path); settingsViewMessages becomes
a thin entry barrel over settingsView/data and settingsView/inbound.
Schemas relocated verbatim; message shapes unchanged.
