+++ src/shared/schemas/progressView/outbound.ts
@@ -1,0 +1,358 @@
+/**
+ * ProgressView outbound message schemas (backend -> frontend): UPDATE_*, SYNC_*,
+ * permission/bypass updates, and the discriminated union they compose into.
+ */
+import { z } from 'zod';
+
+import { PROGRESS_VIEW_COMMANDS } from '@shared/ipc';
+import { GoalStatusSchema } from '@tools/goal/goalMeta';
+
+import { StreamTabIdSchema } from '../identifiers';
+import { StreamLogEntrySchema } from '../log';
+import { AgentOptionDataSchema, ModelOptionDataSchema } from '../mainView';
+import { CompileFailureSchema, OutputFileInfoSchema } from '../output';
+import { InquiryThreadUpdatedEventSchema } from '../inquiry';
+import {
+  AgentProposalPermissionSchema,
+  BashPermissionSchema,
+  ExternalInquiryPermissionSchema,
+  PlanApprovalPermissionSchema,
+  RetryPermissionSchema,
+  ToolEditPermissionSchema,
+  UserQuestionPermissionSchema,
+} from '../prompts';
+import { StreamStatusSchema, StreamTabInfoSchema } from '../stream';
+import {
+  ActiveChildInfoSchema,
+  ConversationProgressSchema,
+  StreamMetadataSchema,
+} from '../streamState';
+import { PlanSchema } from '../plan';
+import { TodoItemSchema } from '../todo';
+import { ContextStateDataSchema } from '../contextManagement';
+import { TokenUsageStatsSchema } from '../usage';
+import { AgentCategoryFilterSchema, ProgressViewPlacementSchema } from './data';
+
+export const UpdateStreamsMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_STREAMS),
+  streams: z.array(StreamTabInfoSchema),
+  activeStream: z.union([StreamTabIdSchema, z.literal('')]),
+  agentFilter: AgentCategoryFilterSchema,
+  streamStates: z.record(z.string(), StreamMetadataSchema).optional(),
+});
+
+export const SetActiveStreamMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.SET_ACTIVE_STREAM),
+  activeStream: z.union([StreamTabIdSchema, z.literal('')]),
+});
+
+export const UpdateConversationProgressMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_CONVERSATION_PROGRESS),
+  stream: StreamTabIdSchema,
+  progress: ConversationProgressSchema,
+});
+
+export const UpdateStreamBadgesMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_STREAM_BADGES),
+  stream: StreamTabIdSchema,
+  activeSubagents: z.array(ActiveChildInfoSchema),
+  finishedSubagentCount: z.number(),
+  activeProcesses: z.array(ActiveChildInfoSchema),
+  finishedProcessCount: z.number(),
+});
+
+export const UpdateProcessOutputMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_PROCESS_OUTPUT),
+  stream: StreamTabIdSchema,
+  executionId: z.string(),
+  stdout: z.string(),
+  stderr: z.string(),
+});
+
+export const UpdateParentStreamMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_PARENT_STREAM),
+  stream: StreamTabIdSchema,
+  parentStreamId: StreamTabIdSchema.nullish(),
+});
+
+export const UpdateStreamDescriptionMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_STREAM_DESCRIPTION),
+  stream: StreamTabIdSchema,
+  description: z.string(),
+});
+
+export const UpdateStreamStatusMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_STREAM_STATUS),
+  stream: StreamTabIdSchema,
+  status: StreamStatusSchema,
+  lastTimestamp: z.number().optional(),
+});
+
+export const LogDeltaMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.LOG_DELTA),
+  streamId: StreamTabIdSchema,
+  entries: z.array(StreamLogEntrySchema),
+  updates: z.array(StreamLogEntrySchema).prefault([]),
+});
+
+export const UpdateFilesMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_FILES),
+  stream: StreamTabIdSchema,
+  rounds: z.record(z.string(), z.array(OutputFileInfoSchema)).optional(),
+  reset: z.boolean().optional(),
+});
+
+export const UpdateMissingOutputsMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_MISSING_OUTPUTS),
+  stream: StreamTabIdSchema,
+  rounds: z.record(z.string(), z.array(z.string())).optional(),
+  reset: z.boolean().optional(),
+});
+
+export const UpdateCompileFailuresMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_COMPILE_FAILURES),
+  stream: StreamTabIdSchema,
+  rounds: z.record(z.string(), z.array(CompileFailureSchema)).optional(),
+  reset: z.boolean().optional(),
+});
+
+export const UpdateTodosMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_TODOS),
+  stream: StreamTabIdSchema,
+  todos: z.array(TodoItemSchema),
+});
+
+export const UpdatePlanMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_PLAN),
+  stream: StreamTabIdSchema,
+  plan: PlanSchema.nullable(),
+});
+
+export const UpdateRunUsageMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_RUN_USAGE),
+  stream: StreamTabIdSchema,
+  runId: z.string(),
+  usage: TokenUsageStatsSchema,
+});
+
+export const UpdateQueuedFollowUpsMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_QUEUED_FOLLOW_UPS),
+  stream: StreamTabIdSchema,
+  messages: z.array(z.string()),
+});
+
+export const SetFollowupOptionsMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.SET_FOLLOWUP_OPTIONS),
+  stream: StreamTabIdSchema,
+  toolUseAgentsData: z.array(AgentOptionDataSchema).optional(),
+  modelOptionsData: z.array(ModelOptionDataSchema).optional(),
+});
+export type SetFollowupOptionsMessage = z.infer<
+  typeof SetFollowupOptionsMessageSchema
+>;
+
+const PermissionKindSchema = z.enum([
+  'toolEdit',
+  'bash',
+  'retry',
+  'proposal',
+  'planApproval',
+  'externalInquiry',
+  'userQuestion',
+]);
+export type ProgressPermissionKind = z.infer<typeof PermissionKindSchema>;
+
+const PermissionPayloadSchema = z.discriminatedUnion('kind', [
+  z.object({
+    kind: z.literal('toolEdit'),
+    data: ToolEditPermissionSchema,
+  }),
+  z.object({
+    kind: z.literal('bash'),
+    data: BashPermissionSchema,
+  }),
+  z.object({
+    kind: z.literal('retry'),
+    data: RetryPermissionSchema,
+  }),
+  z.object({
+    kind: z.literal('proposal'),
+    data: AgentProposalPermissionSchema,
+    modelOptionsData: z.array(ModelOptionDataSchema).optional(),
+    agentOptionsData: z.array(AgentOptionDataSchema).optional(),
+  }),
+  z.object({
+    kind: z.literal('planApproval'),
+    data: PlanApprovalPermissionSchema,
+  }),
+  z.object({
+    kind: z.literal('externalInquiry'),
+    data: ExternalInquiryPermissionSchema,
+  }),
+  z.object({
+    kind: z.literal('userQuestion'),
+    data: UserQuestionPermissionSchema,
+  }),
+]);
+export type PermissionPayload = z.infer<typeof PermissionPayloadSchema>;
+
+export const UpdatePermissionMessageSchema = z.discriminatedUnion('action', [
+  z.object({
+    command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_PERMISSION),
+    action: z.literal('show'),
+    permission: PermissionPayloadSchema,
+  }),
+  z.object({
+    command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_PERMISSION),
+    action: z.literal('resolve'),
+    kind: PermissionKindSchema,
+    id: z.string(),
+  }),
+]);
+export type UpdatePermissionMessage = z.infer<
+  typeof UpdatePermissionMessageSchema
+>;
+
+const BypassTypeSchema = z.enum(['toolEdit', 'superYolo']);
+
+export const UpdateBypassMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_BYPASS),
+  stream: StreamTabIdSchema,
+  type: BypassTypeSchema,
+  bypassActive: z.boolean(),
+});
+
+const FollowUpTextKindSchema = z.enum([
+  'polished',
+  'polishError',
+  'transcribed',
+]);
+
+export const UpdateFollowUpTextMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_FOLLOW_UP_TEXT),
+  stream: StreamTabIdSchema.nullish(),
+  kind: FollowUpTextKindSchema,
+  text: z.string().nullish(),
+  error: z.string().optional(),
+});
+
+const RecordingStatusSchema = z.enum(['started', 'stopped', 'error']);
+
+export const UpdateRecordingMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_RECORDING),
+  status: RecordingStatusSchema,
+  error: z.string().optional(),
+});
+
+export const SyncInquiryThreadsMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.SYNC_INQUIRY_THREADS),
+  threads: z.array(InquiryThreadUpdatedEventSchema),
+});
+
+export const UpdateInquiryThreadMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.UPDATE_INQUIRY_THREAD),
+  thread: InquiryThreadUpdatedEventSchema,
+});
+
+export const SyncStreamContentMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.SYNC_STREAM_CONTENT),
+  stream: z.union([StreamTabIdSchema, z.literal('')]),
+  action: z.enum(['render', 'clear']).optional(),
+  // Workflow flat files (one run per tab)
+  workflowFiles: z.record(z.string(), z.array(OutputFileInfoSchema)).optional(),
+  workflowMissingOutputs: z.record(z.string(), z.array(z.string())).optional(),
+  workflowCompileFailures: z
+    .record(z.string(), z.array(CompileFailureSchema))
+    .optional(),
+  // Per-run usage map — used by both workflow and tool-use so resume
+  // correctly accumulates. Frontend derives sessionUsage as the sum.
+  runUsage: z.record(z.string(), TokenUsageStatsSchema).optional(),
+  contextState: ContextStateDataSchema.optional(),
+  todos: z.array(TodoItemSchema).optional(),
+  plan: PlanSchema.nullable().optional(),
+  queuedFollowUps: z.array(z.string()).optional(),
+  agentCategory: z.string().optional(),
+  // Tab-switch state (R2: replaces separate syncActiveStreamState messages)
+  conversationProgress: ConversationProgressSchema.optional(),
+  badges: z
+    .object({
+      activeSubagents: z.array(ActiveChildInfoSchema),
+      finishedSubagentCount: z.number(),
+      activeProcesses: z.array(ActiveChildInfoSchema),
+      finishedProcessCount: z.number(),
+    })
+    .optional(),
+  parentStreamId: StreamTabIdSchema.optional(),
+  // Toggle bypass state (hydrated on tab switch so toggles display correctly)
+  toolEditBypass: z.boolean().optional(),
+  superYoloBypass: z.boolean().optional(),
+  goalActive: z.boolean().optional(),
+  goalStatus: GoalStatusSchema.optional(),
+  goalObjective: z.string().optional(),
+});
+
+const GoalActiveUpdatedMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.GOAL_ACTIVE_UPDATED),
+  stream: StreamTabIdSchema,
+  active: z.boolean(),
+  status: GoalStatusSchema.optional(),
+  objective: z.string().optional(),
+});
+
+export const ProgressSetThemeMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.THEME_SET),
+  theme: z.enum(['dark', 'light']),
+});
+
+export const ProgressDeleteStreamMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.DELETE_STREAM),
+  stream: StreamTabIdSchema,
+});
+
+export const ProgressDeleteAllMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.DELETE_ALL),
+});
+
+export const SetPlacementMessageSchema = z.object({
+  command: z.literal(PROGRESS_VIEW_COMMANDS.SET_PLACEMENT),
+  placement: ProgressViewPlacementSchema,
+});
+
+export const ProgressViewOutboundMessageSchema = z.discriminatedUnion(
+  'command',
+  [
+    UpdateStreamsMessageSchema,
+    SetActiveStreamMessageSchema,
+    UpdateConversationProgressMessageSchema,
+    UpdateStreamBadgesMessageSchema,
+    UpdateProcessOutputMessageSchema,
+    UpdateParentStreamMessageSchema,
+    UpdateStreamDescriptionMessageSchema,
+    UpdateStreamStatusMessageSchema,
+    LogDeltaMessageSchema,
+    UpdateFilesMessageSchema,
+    UpdateMissingOutputsMessageSchema,
+    UpdateCompileFailuresMessageSchema,
+    UpdateTodosMessageSchema,
+    UpdatePlanMessageSchema,
+    UpdateRunUsageMessageSchema,
+    UpdateQueuedFollowUpsMessageSchema,
+    SetFollowupOptionsMessageSchema,
+    SyncStreamContentMessageSchema,
+    UpdatePermissionMessageSchema,
+    UpdateBypassMessageSchema,
+    GoalActiveUpdatedMessageSchema,
+    UpdateFollowUpTextMessageSchema,
+    UpdateRecordingMessageSchema,
+    SyncInquiryThreadsMessageSchema,
+    UpdateInquiryThreadMessageSchema,
+    SetPlacementMessageSchema,
+    ProgressSetThemeMessageSchema,
+    ProgressDeleteStreamMessageSchema,
+    ProgressDeleteAllMessageSchema,
+  ],
+);
+
+export type ProgressViewOutboundMessage = z.infer<
+  typeof ProgressViewOutboundMessageSchema
+>;
